UPSTREAM: drm: xlnx: zynqmp_dpsub: Support operation without DMA engine
To prepare for usage of the DPSUB as a DisplayPort bridge without
creating a DRM device, make initialization and usage of the DMA engine
optional. The flag that controls this feature is currently hardcoded to
operating with the DMA engine, this will be made dynamic based on the
device tree configuration in a subsequent change.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
(cherry picked from commit 51ae3bd4f0577d250c2b95f58fa93e7937136498)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:268680937
TEST=build test

Backported using forklift.py

Change-Id: Ifd148bcf0d3cc7430f1b4afed0482eb16fba80e9