UPSTREAM: drm: xlnx: zynqmp_dpsub: Decouple DRM device from zynqmp_dpsub
To complete the decoupling of the DRM device from the zynqmp_dpsub,
group all DRM-related structures in a zynqmp_dpsub_drm structure and
allocate it separately from the zynqmp_dpsub. The DRM managed allocation
of the drm_device now doesn't cover the zynqmp_dpsub anymore, so we need
to register a cleanup action to release the zynqmp_dpsub when the
drm_device is released.

The will allow usage of the DisplayPort encoder as a standalone bridge,
without registering a DRM device in this driver.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
(cherry picked from commit d189835fffed7fed4aec15a41eaa27190ad4e04d)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:268680937
TEST=build test

Backported using forklift.py

Change-Id: I76d45fce5ac67a8557fef903b31ae4250c4fad05