CHROMIUM: drm/i915/pxp: handle interrupted PXP TEE messages
If a signal occurs during PXP TEE messages, MEI PXP exits and doesn't
wait for the response. The send still went through though and a response
will come causing the next PXP TEE message to send its own message but
then receive the response from the previously interrupted message and we
end up with all subsequent messages getting out of sync.

Flag whenever a message is interrupted so on the next message we can
drop that message before starting a new send/recv.

Signed-off-by: Juston Li <juston.li@intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Acked-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>

BUG=b:257876317
TEST=emerge-rex sys-kernel/chromeos-kernel-6_1

Change-Id: I8f51e72864616db9307f0f7ec430be6b137a7053
Signed-off-by: Sachin Kumar <sachin.kumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4607174
Reviewed-by: Alan Previn Teres Alexis <alan.previn.teres.alexis@intel.corp-partner.google.com>
Reviewed-by: Juston Li <justonli@google.com>
Reviewed-by: Paz Zcharya <pazz@google.com>
Tested-by: Hasan Abuzaina <hasan.m.abuzaina@intel.corp-partner.google.com>