UPSTREAM: drm/amdkfd: Populate memory info before adding GPU node to topology
The local memory info needs to be fetched before the GPU node is added
to topology. Without this, the sysfs is incorrectly populated and the
size is reported as 0. This was causing rocr tests to fail. This issue
was caused because of a bad merge.

Signed-off-by: Mukul Joshi <mukul.joshi@amd.com>
Reviewed-by: Amber Lin <Amber.Lin@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit d1d22df174ae512c57374f517e346b608f61555c)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:307778453
TEST=cq dry run

Backported using forklift.py

Change-Id: Idc4cc3e3d9e1e73c2da70f57d80511910f6ff3e7