UPSTREAM: drm/i915/selftests: Remove flush_scheduled_work() from live_execlists
There are ongoing efforts to remove usages of flush_scheduled_work() from
drivers in order to avoid several cases of potentential problems when
flushing is done from certain contexts.

Remove the call from the live_execlists selftest. Its purpose was to be
thorough and sync with the execlists capture state handling, but that is
not strictly required for the test to function and can be removed.

Signed-off-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Cc: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Reviewed-by: Nirmoy Das <nirmoy.das@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220630125716.50835-1-tvrtko.ursulin@linux.intel.com
(cherry picked from commit 7d33fd02dd943c7f8003e95930d15d92529fe917)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: If5b967c2cbc2e16557ee595420e9271fa97ee525
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>