BACKPORT: wifi: mac80211: Parse station profile from association response
When processing an association response frame for a Multi-Link
connection, extract the per station profile for each additional
link, and use it for parsing the link elements.

As the Multi-Link element might be fragmented, add support for
reassembling a fragmented element. To simplify memory management
logic, extend 'struct ieee802_11_elems' to hold a scratch buffer,
which is used for the defragmentation. Once an element is
reconstructed in the scratch area, point the corresponding element
pointer to it. Currently only defragmentation of Multi-Link element
and the contained per-STA profile subelement is supported.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 45ebac4f059b92906e7e86dd1a780739f883857c)

Conflicts:
	net/mac80211/util.c
	The variable scratch_len was introduced by this commit. However
	a subsequent upstream commit was cherry picked beforehand to fix
	a security issue:
	ff05d4b45dd89 ("wifi: mac80211: fix MBSSID parsing
	use-after-free")

BUG=b:280636240
TEST=wifi_matfunc and wifi_perf pass on volteer-kernelnext with AX201
TEST=Build can meet rex (with AX211) PDP Enabled Gates
TEST=Basic WiFi7 functionalities like EHT MCS can work on volteer-
     kernelnext with a BE200 swap and interim FW

Change-Id: Idfe383b93cda3af7df159fc169c8f535464e6d82
Signed-off-by: Jun Yu <junyuu@chromium.org>
Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4932773
Reviewed-by: Billy Zhao <billyzhao@chromium.org>