UPSTREAM: dt-bindings: msm: dsi-controller-main: Drop redundant phy-names
Adding in msm8939 which is based msm8916 dtsi I stumbled across a binding
check complaining about the phy name for msm8916 which we were reusing for
msm8939.

The currently inconsistent upstream dtsi naming of "dsi" and "dsi-phy" is
not captured in the yaml for this driver.

The driver however doesn't care about the name of DSI phy, hence the yaml
check is redundant.

Both Krzysztof and Rob suggested we could drop the phy-names entirely if it
really isn't a dependency.

So, drop the inconsistent and unnecessary phy-names field from the yaml.

Signed-off-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
(cherry picked from commit c54155622c29297560a0840640e2d7431add2cb3)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:268680937
TEST=build test

Backported using forklift.py

Change-Id: Ia2407cacb4c8e8222d2635ea79cdadb9b58d493c