UPSTREAM: wifi: mac80211: agg-rx: avoid band check
If the deflink of the station is on 6 GHz, then it won't have HT.
If at the same time we're using MLO, then vif.bss_conf isn't used,
and thus vif.bss_conf.chandef.chan is NULL, causing the code to
crash.

Fix this by just checking for both HT and HE, and refusing the
aggregation session if both are not present. This might be a bit
wrong since it would accept an aggregation session from a peer
that has HE but no HT on 2.4 or 5 GHz, but such a peer shouldn't
exist in the first place, and it probably supports aggregation if
it has HE support.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 2e82be13c6a33cb8befd557f168c92fb363107df)

BUG=b:280636240
TEST=wifi_matfunc and wifi_perf pass on volteer-kernelnext with AX201
TEST=Build can meet rex (with AX211) PDP Enabled Gates
TEST=Basic WiFi7 functionalities like EHT MCS can work on volteer-
     kernelnext with a BE200 swap and interim FW

Change-Id: I3564db14f007448d51dbcd64cd73ba51ea5b9441
Signed-off-by: Jun Yu <junyuu@chromium.org>
Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4932966
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>