UPSTREAM: drm/i915/huc: define gsc-compatible HuC fw for DG2
The fw name is different and we need to record the fact that the blob is
gsc-loaded, so add a new macro to help.

Note: A-step DG2 G10 does not support HuC loading via GSC and would
require a separate firmware to be loaded the legacy way, but that's
not a production stepping so we're not going to bother.

v2: rebase on new fw fetch logic

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Tony Ye <tony.ye@intel.com>
Acked-by: Tony Ye <tony.ye@intel.com>
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220928004145.745803-15-daniele.ceraolospurio@intel.com
(cherry picked from commit a70eebb80022148dfd4d5f60fffd1914ff8e3683)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: I502089e8a72aea792a4f17a003eac7c907769fb3
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>