UPSTREAM: wifi: mac80211: remove support for AddBA with fragmentation
HE added support for dynamic fragmentation inside aggregation
sessions, but no existing driver ever advertises it. Thus,
remove the code for now, it cannot work as-is in MLO. For it
to properly work in MLO, we'd need to validate that the frag
level is identical across all the link bands/iftypes, which
is a good amount of complex code that's just not worth it as
long as no driver has support for it.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit e8d0b807b4a223ae499d43e4baa6c45f946f42d5)

BUG=b:280636240
TEST=wifi_matfunc and wifi_perf pass on volteer-kernelnext with AX201
TEST=Build can meet rex (with AX211) PDP Enabled Gates
TEST=Basic WiFi7 functionalities like EHT MCS can work on volteer-
     kernelnext with a BE200 swap and interim FW

Change-Id: I22bb623522458e7fd31598585eb7f2b5cc14e1a7
Signed-off-by: Jun Yu <junyuu@chromium.org>
Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4932967
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>