UPSTREAM: habanalabs: move reset workqueue to be under hl_device
'struct hl_device_reset_work' is used as a wrapper for the reset work
and its parameters, including the reset workqueue on which it runs.
In a future commit, another reset related work with similar parameters
is going to be added, but it won't use the reset workqueue.

As in any case there is a single reset workqueue, and to allow the resue
of this structure, move the reset workqueue to 'struct hl_device'.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit 1eebb259290b1be5398fec953bdd7923a5cbf33e)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:300094055
TEST=build test

Backported using forklift.py

Change-Id: I6b029c887c8aff1ef2e0c844990c01ede6802661