UPSTREAM: phy: dphy: add support to calculate the timing based on hs_clk_rate
For MIPI-CSI sender use-case it is common to specify the allowed
link-frequencies which should be used for the MIPI link and is
half the hs-clock rate.

This commit adds a helper to calculate the D-PHY timing based on the
hs-clock rate so we don't need to calculate the timings within the
driver.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Acked-by: Vinod Koul <vkoul@kernel.org>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
(cherry picked from commit 22168675bae75e158c459ef2dee3b6ebd52a80ed)

BUG=b:307778453
TEST=cq dry run

Change-Id: I509572b049d73ae8fde8a4b8a61c341063dc5a6a
Signed-off-by: Drew Davenport <ddavenport@chromium.org>