UPSTREAM: drm/i915/guc: Handle save/restore of MCR registers explicitly
MCR registers can be placed on the GuC's save/restore list, but at the
moment they are always handled in a multicast manner (i.e., the GuC
reads one instance to save the value and then does a multicast write to
restore that single value to all instances).  In the future the GuC will
probably give us an alternate interface to do unicast per-instance
save/restore operations, so we should be very clear about which
registers on the list are MCR registers (and in the future which
save/restore behavior we want for them).

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Balasubramani Vivekanandan <balasubramani.vivekanandan@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221014230239.1023689-11-matthew.d.roper@intel.com
(cherry picked from commit cf35f6afb92643633f4ecbb386ab8a572cca0386)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: Iae46f3768e9b9770328ef2a2b605f72d642f525c
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>