BACKPORT/UPSTREAM: drm: Remove drm_mode_config::fb_base
The fb_base in struct drm_mode_config has been unused for a long time.
Some drivers set it and some don't leading to a very confusing state
where the variable can't be relied upon, because there's no indication
as to which driver sets it and which doesn't.

The only usage of fb_base is internal to two drivers so instead of trying
to force it into all the drivers to get it into a coherent state
completely remove it.

Change-Id: I3301512302166a9baa8af350cf079663394a1990
Signed-off-by: Zack Rusin <zackr@vmware.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Thomas Zimmermann <tzimemrmann@suse.de>
Acked-by: Christian König <christian.koenig@amd.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221019024401.394617-1-zack@kde.org
(cherry picked from commit 7c99616e3fe7f35fe25bf6f5797267da29b4751e)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:268680937
TEST=build test

Backported using forklift.py

Change-Id: I19cd769d0108e5063773b18d4079781a8a9a37e2