UPSTREAM: HID: Recognize "Digitizer" as a valid input application
"Digitizer" is a generic usage that may be used by various devices but
which is particularly used by non-display pen tablets. This patch adds the
usage to the list of values matched by the IS_INPUT_APPLICATION() macro
that determines if an input device should be allocated or not.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>
Link: https://lore.kernel.org/r/20221019151832.44522-1-jason.gerecke@wacom.com
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
(cherry picked from commit a0f5276716c80415230b47e321e1c46357d70993)

BUG=b:268223063
TEST=CQ passes

Change-Id: Ib014ed5ed15feb1a809cdc06ecc434fe444a4afc
Signed-off-by: Jingyuan Liang <jingyliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4442746
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>