UPSTREAM: ALSA/ASoC: hda: move SPIB/DRMS functionality from ext layer
The SPIB and DRMS capabilities are orthogonal to the DSP enablement
and can be used whether the stream is coupled or not.

The existing code partitioning makes limited sense, the capabilities
are parsed at the sound/hda level but helpers are located in
sound/hda/ext.

This patch moves all the SPIB/DRMS functionality to the sound/hda
layer. This reduces the complexity of the sound/hda/ext layer which is
now limited to handling the multi-link extensions and stream
coupling/decoupling helpers.

Note that this is an iso-functionality code move and rename, the
HDaudio legacy driver would need additional changes to make use of
these capabilities.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Link: https://lore.kernel.org/r/20221019162115.185917-11-pierre-louis.bossart@linux.intel.com
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 6258234129b013c534fa10abaf08751b2401b22b)

BUG=b:266649126
TEST=Test Audio use cases

Change-Id: Ie77aec84c5257736f5f37407328ad804d65b8f93
Signed-off-by: Subbarao Aketi <aketix.subbarao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4280965
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Commit-Queue: Terry Cheong <htcheong@chromium.org>
Tested-by: Terry Cheong <htcheong@chromium.org>