UPSTREAM: drm/amd/display: Stop waiting for vblank during pipe programming
[why and how]
This line was originally removed for a compliance issue, but then
reverted as it caused a fringe underflow case.

However, the addition of this line caused a underflow regression
when subVP is on, and it needs to be removed again. We plan to
fix subvp underflow and then re-add in this line. After that,
we will investigate what to do next for the compliance issue.

Reviewed-by: Alvin Lee <Alvin.Lee2@amd.com>
Acked-by: Alex Hung <alex.hung@amd.com>
Signed-off-by: Martin Leung <Martin.Leung@amd.com>
Tested-by: Mark Broadworth <mark.broadworth@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 935e35d8dbde76a02d00290cf4d9eb420514eaf3)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:268680937
TEST=build test

Backported using forklift.py

Change-Id: If1fc6e9581cfe7028f7f133f31dc896eb9b1717b