UPSTREAM: linux/container_of.h: Warn about loss of constness
container_of() casts the original type to another which leads to the loss
of the const qualifier if it is not specified in the caller-provided type.
This easily leads to container_of() returning a non-const pointer to a
const struct which the C compiler does not warn about.

Acked-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Link: https://lore.kernel.org/r/20221024111627.75183-1-sakari.ailus@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 7376e561fd2e017e9a53f975209777234b8b434e)

BUG=b:307778453
TEST=cq dry run

Change-Id: Ia1092d8851594a900319f864980669727f593dbf
Signed-off-by: Drew Davenport <ddavenport@chromium.org>