UPSTREAM: resource: Convert DEFINE_RES_NAMED() to be compound literal
Currently DEFINE_RES_NAMED() can only be used to fill the static data.
In some cases it would be convenient to use it as right value in the
assignment operation. But it can't be done as is, because compiler has
no clue about the data layout. Converting it to be a compound literal
allows the above mentioned usage.

Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
Reviewed-by: Rafael J. Wysocki <rafael@kernel.org>
Link: https://lore.kernel.org/r/20221109155618.42276-2-andriy.shevchenko@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 52c4d11f1dce60453ab2a75fd7103118cedb2b58)

BUG=b:280653303
TEST=build test

Change-Id: I562c0410e51c9076d4cc86b49eda743462dd14be
Signed-off-by: Drew Davenport <ddavenport@chromium.org>