BACKPORT: fuse: add request extension
Will need to add supplementary groups to create messages, so add the
general concept of a request extension.  A request extension is appended to
the end of the main request.  It has a header indicating the size and type
of the extension.

The create security context (fuse_secctx_*) is similar to the generic
request extension, so include that as well in a backward compatible manner.

Add the total extension length to the request header.  The offset of the
extension block within the request can be calculated by:

  inh->len - inh->total_extlen * 8

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Link: https://lore.kernel.org/all/20221110142020.191487-2-mszeredi@redhat.com/
(cherry picked from commit 15d937d7ca8c55d2b0ce9116e20c780fdd0b67cc)

Downstream changes:
include/uapi/linux/fuse.h
Version lable 7.38 added in FUSE protocol changelog.

BUG=b:307851138
TEST=CQ

Change-Id: I483ac2c0dbe9e521f987da9d073e1136486b2f40
Signed-off-by: Yuan Yao <yuanyaogoog@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4978298
Reviewed-by: Takaya Saeki <takayas@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>