UPSTREAM: habanalabs: print context refcount value if hard reset fails
Failing to kill a user process during a hard reset can be due to a
reference to the user context which isn't released.
To make it easier to understand if this the reason for the failure and
not something else, add a print of the context refcount value.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit 408c46bd6eb7a4e2fb9fd686218e4a13b9de844c)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:300094055
TEST=build test

Backported using forklift.py

Change-Id: I0cc7a37a0afb476cb97b96701a67d9372d9a68ec