FROMGIT: drm/i915: Finish the LUT state checker
We have full readout now for all platforms (sans the icl+
multi-segment readout hw fail), so hook up the LUT state
checker for everyone.

We add a new vfunc for this since different platforms need
to handle the details a bit differently.

The implementation is rather repetitive in places. Probably
we want to think of a more declarative approach for the
LUT precision/etc. stuff in the future...

Note that we're currently missing readout for c8_planes, so
we'll have to skip the state check in that case.

v2: Fix readout for C8 use cases
v3: Skip C8 entirely due to lack of c8_planes readout
    Add ilk_has_pre_csc_lut() helper and use other such helpers

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221114153732.11773-16-ville.syrjala@linux.intel.com
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
(cherry picked from commit 96ffd0cda60abfbb39246ed2fb0d344b57289904
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: I143b7ee8d88fb5cde7f20c24761ae5544663bbfa
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>