FROMGIT: drm/i915: Use gamma LUT for RGB limited range compression
On hsw+ and glk class hardware we current make a mess of
things when we have to both generate limited range output
and use the hw gamma LUT. Since we do the range compression
using the pipe CSC unit (which is situated before the gamma
LUT in the pipe) we are in fact applying the gamma to the
limited range data instead of the full range data as the
user intended.

We can work around this by applying the range compression
via the gamma LUT instead of using the pipe CSC for it.
Fairly easy to do now that we have the internal post_csc_lut
attachment point where we can stick our new cooked LUT.

On hsw+ this only needs to be done when using the split
gamma mode or when the ctm is enabled since otherwise we can
simply reorder the LUT vs. CSC. On glk we need to do this any
time a gamma LUT is used since no reordering is possible.
We do lose a bit of coverage in intel_color_assert_luts(),
but so be it.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221114153732.11773-19-ville.syrjala@linux.intel.com
Reviewed-by: Uma Shankar <uma.shankar@intel.com>
(cherry picked from commit 07fc6a7b1a03934cd564ce8492bb6e7a49a12751
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: Ic9c5e43e4ae3e673c48a9452956e4faaf38ec35f
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>