UPSTREAM: thermal: intel: Prevent accidental clearing of HFI status
When there is a package thermal interrupt with PROCHOT log, it will be
processed and cleared. It is possible that there is an active HFI event
status, which is about to get processed or getting processed. While
clearing PROCHOT log bit, it will also clear HFI status bit. This means
that hardware is free to update HFI memory.

When clearing a package thermal interrupt, some processors will generate
a "general protection fault" when any of the read only bit is set to 1.

The driver maintains a mask of all read-write bits which can be set.

This mask doesn't include HFI status bit. This bit will also be cleared,
as it will be assumed read-only bit. So, add HFI status bit 26 to the
mask.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Reviewed-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 6fe1e64b60269aa58fa00568807738025ae3bd05)

BUG=b:287305592
TEST=emerge-rex sys-kernel/chromeos-kernel-6_1

Change-Id: Id664d7c788d558d677df8a37ba472eae7f305f8f
Signed-off-by: Sachin Kumar <sachin.kumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4625821
Tested-by: Adeel Arshad <adeel.arshad@intel.corp-partner.google.com>
Tested-by: Waheem Md <waheem.md@intel.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Paz Zcharya <pazz@google.com>
Commit-Queue: Paz Zcharya <pazz@google.com>