UPSTREAM: habanalabs: make print of engines idle mask more readable
The engines idle mask was increased to be an array of 4 u64 entries.
To make the print of this mask more readable, remove the "0x" prefix,
and zero-pad each u64 to 16 bytes if either it isn't zero or if any of
the higher-order u64's is not zero.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit 1b18cf33d6ce63a9f5fe3764d7b20c4738dd1245)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:300094055
TEST=build test

Backported using forklift.py

Change-Id: I7411d967905e7ea63c7bcb99aa280248160f9b03