UPSTREAM: wifi: mac80211: Fix for Rx fragmented action frames
The ieee80211_accept_frame() function performs a number of early checks
to decide whether or not further processing needs to be done on a frame.
One of those checks is the ieee80211_is_robust_mgmt_frame() function.
It requires to peek into the frame payload, but because defragmentation
does not occur until later on in the receive path, this peek is invalid
for any fragment other than the first one. Also, in this scenario there
is no STA and so the fragmented frame will be dropped later on in the
process and will not reach the upper stack. This can happen with large
action frames at low rates, for example, we see issues with DPP on S1G.

This change will only check if the frame is robust if it's the first
fragment. Invalid fragmented packets will be discarded later after
defragmentation is completed.

Signed-off-by: Gilad Itzkovitch <gilad.itzkovitch@morsemicro.com>
Link: https://lore.kernel.org/r/20221124005336.1618411-1-gilad.itzkovitch@morsemicro.com
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit e6f5dcb7ec9badd9500f64b087f7243c37300d63)

BUG=b:280636240
TEST=wifi_matfunc and wifi_perf pass on volteer-kernelnext with AX201
TEST=Build can meet rex (with AX211) PDP Enabled Gates
TEST=Basic WiFi7 functionalities like EHT MCS can work on volteer-
     kernelnext with a BE200 swap and interim FW

Change-Id: I53ba8c81c474146b7fc9daf4f6bf810e212e22d6
Signed-off-by: Jun Yu <junyuu@chromium.org>
Signed-off-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4933022
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>