CHROMIUM: iwl7000: improve debug prints in iwl_read_ppag_table()
Add prints of the ETSI and CHINA bits.
Check if need to mask the CHINA bit (when FW doesn't support it) in a
separate flow so it will be easier to follow the conditions.
Add a message for the cmd version.
Change existing messages to describe in what revision the FW supports
(and not what is the cmd ver).

Signed-off-by: Alon Giladi <alon.giladi@intel.com>
Change-Id: I6f1a4b7dfa5cfd9b4f4a4b5bb5567849e629ae96
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: c939fafbd0c0f7ff2c2a1790b6e4cffaa2ec7bd2