BACKPORT: FROMGIT: drm/i915/gsc: Skip the version check when fetching the GSC FW
The current exectation from the FW side is that the driver will query
the GSC FW version after the FW is loaded, similarly to what the mei
driver does on DG2. However, we're discussing with the FW team if there
is a way to extract the version from the bin file before loading, so we
can keep the code the same as for older FWs.

Since the GSC FW version is not currently required for functionality and
is only needed for debug purposes, we can skip the FW version for now at
fetch time and add it later on when we've agreed on the approach.

v2: rebased on uc_fw version struct changes.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Cc: Alan Previn <alan.previn.teres.alexis@intel.com>
Cc: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Rodrigo Vivi <rodrigo.vivi@intel.com> #v1
Reviewed-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20221208200521.2928378-3-daniele.ceraolospurio@intel.com
(cherry picked from commit f0ce5178a38918adfef728f43dd544616b05e3e3
 https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

Conflicts:
   drivers/gpu/drm/i915/gt/uc/intel_uc_fw.c

BUG=b:265114629
TEST=i915 driver loaded successfully

Change-Id: Ie2469c70aa4f06fa95312ae361446b7d302a1592
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>