UPSTREAM: habanalabs: fix dma-buf release handling if dma_buf_fd() fails
The dma-buf private object is freed if a call to dma_buf_fd() fails,
and because a file was already associated with the dma-buf in
dma_buf_export(), the release op will be called and will use this
object.

Mark the 'priv' field as NULL in this case, and avoid accessing it from
the release op.

Signed-off-by: Tomer Tayar <ttayar@habana.ai>
Reviewed-by: Oded Gabbay <ogabbay@kernel.org>
Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit 9b3d9f917f81aed559d19377c11d6fac3db89fd7)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:300094055
TEST=build test

Backported using forklift.py

Change-Id: Ie23420fc9aa7f4f3c91b1f9959e8a971a955402c