UPSTREAM: habanalabs: move some prints to debug level
When entering an IOCTL, the driver prints a message in case device is
not operational. This message should be printed in debug level as
it can spam the kernel log and it is not an error.

Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit e1e8e7472b8c647c4e117191d6b44b852c90cd24)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:300094055
TEST=build test

Backported using forklift.py

Change-Id: Id46febea9d08fe6245312f87b9e4ba2a7d8735d2