UPSTREAM: thunderbolt: Reset USB4 v2 host router
USB4 v2 added a bit that can be used to reset the host router so we use
this to trigger reset when the driver probes. This will reset the
already connected topology as well but doing this simplifies things a
lot if for instance the link is already set to asymmetric. We also add
a module parameter to prevent this in case of problems.

While there rename the REG_HOP_COUNT to REG_CAPS to match the USB4 spec
naming better.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 0fc70886569c8459c4494ba9ef8c4ef34b81e781)

BUG=b:294499201
TEST=TBT3/4 functional tests to ensure backward compatibility

Change-Id: Ie745e3e6ee623187c30a5bd4891f44b1812c3162
Signed-off-by: Jignesh Patel <jignesh1.patel@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4997582
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>