BACKPORT: habanalabs: move driver to accel subsystem
Now that we have a subsystem for compute accelerators, move the
habanalabs driver to it.

This patch only moves the files and fixes the Makefiles. Future
patches will change the existing code to register to the accel
subsystem and expose the accel device char files instead of the
habanalabs device char files.

Update the MAINTAINERS file to reflect this change.

Signed-off-by: Oded Gabbay <ogabbay@kernel.org>
(cherry picked from commit e65e175b07bef5974045cc42238de99057669ca7)

Conflicts:
   drivers/Makefile
   drivers/accel/Makefile

BUG=b:300094055
TEST=build test

Change-Id: I1c1aa469c432be5b7058b447a73675c79e9205f0
Signed-off-by: Drew Davenport <ddavenport@chromium.org>